备份

+++ Assistant/app/src/main/java/com/example/zwq/assistant/Activity/MeFragment.java
@@ -161,7 +161,7 @@ public class MeFragment extends BaseFragment {
                             tvPhone.setText(userHttpResult.getData().getPhone());
                             tvClass.setText(userHttpResult.getData().getClassName());
                             tvParentPho.setText(userHttpResult.getData().getParentPho());
-                            int number = userHttpResult.getData().getNumber();
+                            double number = userHttpResult.getData().getNumber();
                             tvNumber.setText(String.valueOf(number));
                             int sex = userHttpResult.getData().getSex();
                             if (sex == 1){
